Change operator to filter by keywords
The default operator used by Elasticsearch in the query to filter
documents by keyword is an "OR". This means that any documents
containing one or more of the keywords is returned. For now, we want to
filter only documents which contains all the keywords. This commit
change the default operator to "AND" to achieve this goal.

Signed-off-by: José Guilherme Vanz <jvanz@jvanz.com>