Edition and extra flag in the response
Adds two more optional fields in the response model to send the edition
number and a flag to show if the gazette is extra or not. Furthermore,
the endpoint has two more additional configuration to remote fields with
None value or if they are not set. Thus, the edition and extra flag is
removed when necessary.

Signed-off-by: José Guilherme Vanz <jvanz@jvanz.com>