Return number of item and change pagination
Adds the total number of item returned by the search in the API
response. Change also the pagination to be based on a offset instead of
the page number. Thus, the caller can define how many item should be
skipped in the response. Allowing the user to control the pagination
in the client side.

Signed-off-by: José Guilherme Vanz <jvanz@jvanz.com>