Cache podman layers in GitHub workflows (#12)
I enabled GitHub caching for podman's layer cache folder
(`~/.local/share/containers/`) and reordered the commands in the
Dockerfile to optimize the cache usage by keeping the things that change
the most (i.e. the code) in the last build steps.

GitHub will automatically delete the cache if it's unused after a few
days.